Add jest for testing API
- Add jest and related packages
- Add simple test case for `/banner/:word` endpoint
