+++ api/server.js
@@ -1,33 +1,5 @@
-const express = require('express')
-const cors = require('cors')
-const app = express()
-
-const { getSingleWordHandler } = require('./src/handlers/words')
-const { getImageHandler } = require('./src/handlers/banner')
-const { getSummaryHandler } = require('./src/handlers/summary')
-const { getSuggestions } = require('./src/handlers/suggestions')
-
-const { validateGetSuggestions } = require('./src/validators/suggestions')
-
+const app = require('./src/app')
 const PORT = process.env.PORT ?? 8080
-
-app.use(
-  cors({
-    origin: [/\.tudien\.io$/],
-  }),
-)
-
-app.use(express.json())
-
-app.disable('x-powered-by')
-
-app.get('/words/:word', getSingleWordHandler)
-
-app.get('/banner/:word', getImageHandler)
-
-app.get('/summary', getSummaryHandler)
-
-app.get('/suggestions/:word', validateGetSuggestions, getSuggestions)
 
 app.listen(PORT, function (err) {
   if (err) console.log(err)
